# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Client API",
        "description": "API documentation for Imperial Client Mobile & Web Applications",
        "contact": {
            "email": "support@imperial.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List authenticated client bookings",
                "operationId": "bfcd02212e1e0f38f278be30ad3927aa",
                "parameters": [
                    {
                        "name": "club_id",
                        "in": "query",
                        "description": "Filter bookings by a specific club",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "Filter bookings by a specific event",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "table_id",
                        "in": "query",
                        "description": "Filter bookings by a specific table",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking_date",
                        "in": "query",
                        "description": "Filter bookings by a specific date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking request",
                "operationId": "f565d3c9cc8ac61e18bdf09d07a34123",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_date",
                                    "table_id",
                                    "club_id",
                                    "guest_ids"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "booking_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-23"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "2:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "22:00"
                                    },
                                    "discount_source": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "discount_code": {
                                        "type": "string",
                                        "example": "ABCF123"
                                    },
                                    "discount_note": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Say something related to booking or instruction something"
                                    },
                                    "guest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "booking": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or table occupied"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking details",
                "operationId": "3daad65b4c4d8c9b5f6f4459cece3b58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings/{id}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "c48b00af953cae001a48d12447a732d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking cancelled successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Checked in bookings cannot be cancelled"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum Bearer token",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}